Revert "Add the directory appinsights"
This reverts commit 59a088a22bf1f9a3d904fe1921392ee89d48f2fb.
