--- appinsights/appinsights.phpproj
@@ -1,29 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Name>appinsights</Name>
-    <ProjectGuid>a9bb5047-f989-4d4c-a696-d6b351871df1</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <RootNamespace>appinsightsaddon</RootNamespace>
-    <ProjectTypeGuids>{A0786B88-2ADB-4C21-ABE8-AA2D79766269}</ProjectTypeGuids>
-    <AssemblyName>appinsightsaddon</AssemblyName>
-    <ProjectHome />
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
-    <IncludeDebugInformation>true</IncludeDebugInformation>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
-    <IncludeDebugInformation>false</IncludeDebugInformation>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="blocks\appinsights\add.php" />
-    <Compile Include="blocks\appinsights\controller.php" />
-    <Compile Include="blocks\appinsights\edit.php" />
-    <Compile Include="Controller.php" />
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="blocks\" />
-    <Folder Include="blocks\appinsights\" />
-  </ItemGroup>
-</Project>
