--- appinsights/blocks/appinsights/add.php
@@ -1,3 +1,0 @@
-<?php
-
-?>
