--- appinsights/blocks/appinsights/edit.php
@@ -1,3 +1,0 @@
-<?php
-
-?>
