--- appinsights/composer.json
@@ -1,14 +1,0 @@
-{
-    "require": {
-        "php": ">=5.4.2",
-        "microsoft/application-insights": ">=0.2.4"
-    },
-    "replace": { 
-        "guzzlehttp/guzzle": "~5.0" 
-    },
-    "autoload": {
-        "psr-4": {
-            "ApplicationInsights\\Concrete\\": "src/"
-        }
-    }
-}
